Add const test.
Not perfect. But it works.

The SUT should not read the message for the decode test, but rather
print the decoded string using the const value from the schema.

Likewise, it should set the msg field of the SimpleTest struct to that
of the const value explicitly, rather sending a default (empty) struct
value.